Changing price to float

+++ RequestsResponses/EditProductRequest.cs
@@ -13,7 +13,7 @@ namespace personalProjectAPI.RequestsResponses
 
         public string? Description { get; set; }
 
-        public double? Price { get; set; }
+        public float? Price { get; set; }
 
         public string? ImageUrl { get; set; }
     }
